+++ navigation/core/lib/repository/location_repository.dart
@@ -1,0 +1,5 @@
+import 'package:navigation_core/model/coordinate_model.dart';
+
+abstract class LocationRepository {
+  Future<CoordinateModel?> getCurrentLocation();
+}