+++ navigation/impl_foss/lib/di/navigation_impl_module_factory.dart
@@ -1,0 +1,14 @@
+import 'package:navigation_core/repository/location_repository.dart';
+import 'package:navigation_core/repository/route_repository.dart';
+import 'package:navigation_impl/repository/foss_route_repository.dart';
+import 'package:navigation_impl/repository/platform_location_repository.dart';
+
+class NavigationImplModuleFactory {
+  static RouteRepository createRouteRepository() {
+    return FossRouteRepository();
+  }
+
+  static LocationRepository createLocationRepository() {
+    return PlatformLocationRepository();
+  }
+}