+++ station/lib/repository/dto/station_dto.dart
@@ -1,5 +1,5 @@
 import 'package:collection/collection.dart';
-import 'package:navigation/coordinate_model.dart';
+import 'package:navigation/navigation.dart';
 import 'package:station/model/currency_model.dart';
 import 'package:station/model/station_model.dart';